+++ database/migrations/2024_09_24_101113_create_annonces_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->dateTime('date_depart');
             $table->dateTime('date_arrivee');
             $table->text('description');
-            $table->float('prix_du_kilo');
+            $table->decimal('prix_du_kilo');
             $table->string('origin');
             $table->string('destination');
             $table->timestamps();
